1.8
Modificaciones varias

+++ Veterinaria/src/veterinaria/AccesoADatos/ClienteData.java
@@ -119,7 +119,7 @@ public class ClienteData {
                 cliente.setApellido(rs.getString("apellido"));
                 cliente.setNombre(rs.getString("nombre"));
                 cliente.setDireccion(rs.getString("direccion"));
-                cliente.setTelefono(rs.getInt("telefono"));
+                cliente.setTelefono(rs.getLong("telefono"));
                 cliente.setPersonaAlternativa(rs.getString("personaAlternativa"));
                 cliente.setEstado(rs.getBoolean("estado"));
             }
